refactor(InputContainer): 파일명 수정 및 count 컴포넌트 분리
