+++ frontend/src/components/common/Count/Count.styled.ts
@@ -1,18 +1,5 @@
 import styled from "@emotion/styled";
 
-export const InputField = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 0.8rem;
-
-  width: 100%;
-`;
-
-export const Label = styled.label`
-  ${({ theme }) => theme.typography.mobile.bodyBold};
-  color: ${(props) => props.theme.colors.text.primary};
-`;
-
 export const CountWrapper = styled.div`
   display: flex;
   justify-content: flex-end;
