+++ frontend/src/components/common/Count/Count.tsx
@@ -1,0 +1,16 @@
+import * as S from "./Count.styled";
+
+interface CountProps {
+  count: number;
+  maxCount: number;
+}
+
+const Count = ({ count, maxCount }: CountProps) => {
+  return (
+    <S.CountWrapper>
+      <S.Count>{`${count}/${maxCount}`}</S.Count>
+    </S.CountWrapper>
+  );
+};
+
+export default Count;
